Se crea controller de Usuario

+++ api/inmoboscoapi/src/main/java/com/salesianostriana/pdam/inmoboscoapi/services/UserService.java
@@ -9,6 +9,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import java.text.SimpleDateFormat;
+import java.time.LocalDate;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -25,30 +27,37 @@ public class UserService {
                 .firstname(createUserRequest.getFirstname())
                 .lastname(createUserRequest.getLastname())
                 .username(createUserRequest.getUsername())
-                .password(createUserRequest.getPassword())
-                .phone(createUserRequest.getPhoneNumber())
+                .password(passwordEncoder.encode(createUserRequest.getPassword()))
+                .phoneNumber(createUserRequest.getPhoneNumber())
                 .dni(createUserRequest.getDni())
                 .avatar(createUserRequest.getAvatar())
                 .email(createUserRequest.getEmail())
-                .birthdate(createUserRequest.getBirthdate())
+                .birthdate(LocalDate.parse(createUserRequest.getBirthdate()))
                 .rol(roles)
                 .build();
         return userRepository.save(user);
     }
 
+    public User createUserWithWorkerRole(CreateUserRequest createUserRequest){
+        return createUser(createUserRequest, EnumSet.of(UserRole.WORKER));
+    }
+
+    public User createUserWithOwnerRole(CreateUserRequest createUserRequest){
+        return createUser(createUserRequest, EnumSet.of(UserRole.OWNER));
+    }
+    public User createUserWithUserRole(CreateUserRequest createUserRequest){
+        return createUser(createUserRequest, EnumSet.of(UserRole.USER));
+    }
+
     public List<CreateUserResponse> findAllUsers() {
         List<User> data = userRepository.findAll();
 
-        return data.stream()
-                .map(CreateUserResponse::createUserResponseFromUser)
-                .collect(Collectors.toList());
+        return data.stream().map(CreateUserResponse::createUserResponseFromUser).collect(Collectors.toList());
     }
 
     public CreateUserResponse findUserById(UUID id) {
         /*TODO: ESTE METODO HAY QUE MODIFICARLO CUANDO NOS PONGAMOS CON LA GESTION DE ERRORES*/
-        return userRepository.findById(id)
-                .map(CreateUserResponse::createUserResponseFromUser)
-                .orElseThrow();
+        return userRepository.findById(id).map(CreateUserResponse::createUserResponseFromUser).orElseThrow();
     }
 
     public Optional<User> findUserByUsername(String username) {
@@ -60,41 +69,23 @@ public class UserService {
         /*TODO: ESTE METODO HAY QUE MODIFICARLO CUANDO NOS PONGAMOS CON LA GESTION DE ERRORES*/
 
         /*TODO:PRGUNTAR SI AQUI MERECE LA PENA LA ENTIDAD AL COMPLETO EN VEZ DEL DTOs*/
-        return userRepository.findFirstByUsername(username)
-                .map(CreateUserResponse::createUserResponseFromUser)
-                .orElseThrow();
+        return userRepository.findFirstByUsername(username).map(CreateUserResponse::createUserResponseFromUser).orElseThrow();
     }
 
     public Optional<User> editPassword(UUID userId, String newPassword) {
-        return userRepository.findById(userId)
-                .map(u -> {
-                    u.setPassword(passwordEncoder.encode(newPassword));
-                    return userRepository.save(u);
-                }).or(Optional::empty);
+        return userRepository.findById(userId).map(u -> {
+            u.setPassword(passwordEncoder.encode(newPassword));
+            return userRepository.save(u);
+        }).or(Optional::empty);
 
     }
 
     public void deleteUserByID(UUID id) {
         /*TODO: ESTE METODO HAY QUE MODIFICARLO CUANDO NOS PONGAMOS CON LA GESTION DE ERRORES*/
-        if (userRepository.existsById(id))
-            userRepository.deleteById(id);
+        if (userRepository.existsById(id)) userRepository.deleteById(id);
     }
-
-
-
 
 }
 
 
 
-
-
-    /*
-    public User createUserWithOwnerRole(CreateUserRequest createUserRequest){
-        return createUser(createUserRequest, Set.of(UserRole.OWNER));
-    }
-    */
-  /*  public User createUserWithWorkerRole(CreateUserRequest createUserRequest){
-        return createUser(createUserRequest, Set.of(UserRole.WORKER));
-    }
-*/
